requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.50.0
PyYAML>=5.3.0
lmdb>=0.99
requests>=2.25.0
scikit-image>=0.18.0
scipy>=1.5.0
h5py>=2.10.0
python-dotenv>=0.19.0
